scripts/test/requirements.txt
Flask==1.0.2
requests==2.24.0
python-dotenv==0.19.0
